scripts/app-deploy/requirements.txt
python-dotenv
azure-identity
requests
